wip: Support for generating index files
Create a gha that can generate an index for every query in the src/sql
folder. The gha will check if the version of IDC is out of date.

If so, the queries will be updated, and run with bigquery to generate csv
and parquet files in root directory. The cmake lists file is edited to pick
up the csv and parquet files generated by bq and package them.

Moreover gitignore is edited to have a pattern for ignoring service account files.

Lastly, if queries are updated, a pull request will be created to update the
queries to be in sync with latest idc release.