enh: add sm_index query (#27)
Re #15

Adding query that replicates the content of the existing `bigquery-public-data.idc_v18.dicom_metadata_curated_series_level`
BigQuery table (with the exception of the `Modality` column). This
is non-controversial, since we have already discussed the content of
that table, it already flattened, and it is sufficient to address the
immediate needs of the team and should be sufficient to meet the needs
of the user in https://discourse.canceridc.dev/t/metadata-information-for-pathology-images/595.