ENH: added clinical_index (#35)
Originally, this was populated directly from Parquet AWS content in idc-index,
however, it contains columns that are of no direct use that may confuse the user.
Further, accessing this index from AWS introduced dependency on the AWS Parquet
export being up to date. Since this index is quite small, it is best to replace it with a query.